feat: refactorizar la estructura de navegación creando el componente AppNavigator
