+++ src/navigation/AppNavigator.js
@@ -1,0 +1,19 @@
+import { NavigationContainer } from "@react-navigation/native";
+import { createNativeStackNavigator } from "@react-navigation/native-stack";
+import AuthScreen from "../screens/AuthScreen";
+
+const Stack = createNativeStackNavigator();
+
+export default function AppNavigator() {
+  return (
+    <NavigationContainer>
+      <Stack.Navigator
+        initialRouteName="Auth"
+        screenOptions={{ headerShown: false }}
+      >
+        <Stack.Screen name="Auth" component={AuthScreen} />
+        {/* agrega más pantallas aquí */}
+      </Stack.Navigator>
+    </NavigationContainer>
+  );
+}
